pyproject.toml
[tool.poetry]
name = "liquid-level-drone"
version = "0.1.0"
description = "This script enables the detection of liquid levels (empty, half, full) using drone technology and YOLOv5. It can be applied in various scenarios, such as beverage packaging factories or chemical facilities, to monitor and manage liquid levels effectively."
authors = ["Chen Feng Tsai <rich.tsai1103@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
azure-cognitiveservices-speech = "^1.32.1"
djitellopy = "^2.5.0"



[tool.poetry.group.dev.dependencies]
gitpython = ">=3.1.30"
matplotlib = ">=3.3"
numpy = ">=1.22.2"
opencv-python = ">=4.1.1"
pillow = ">=7.1.2"
psutil = "^5.9.6"
resources = "^0.0.1"
pyyaml = ">=5.3.1"
requests = ">=2.23.0"
scipy = ">=1.4.1"
thop = ">=0.1.1"
flops = "^0.1.4"
torch = ">=1.8.0"
see = "^1.4.1"
torchvision = ">=0.9.0"
tqdm = ">=4.64.0"
ultralytics = ">=8.0.147"
pandas = ">=1.1.4"
seaborn = ">=0.11.0"
setuptools = ">=65.5.1"
fix = "^1.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
